requirements.txt

streamlit==1.37.1
python-pptx==0.6.23
Pillow==10.4.0
yake==0.4.8
